+++ mathics/eval/image.py
